ASSIGNMENT UPLOAD
